+++ cmd2/__init__.py
@@ -14,6 +14,7 @@ from . import (
 from .argparse_completer import set_default_ap_completer_type
 from .argparse_custom import (
     Cmd2ArgumentParser,
+    TextGroup,
     register_argparse_argument_parameter,
     set_default_argument_parser_type,
 )
@@ -59,6 +60,7 @@ __all__: list[str] = [  # noqa: RUF022
     'DEFAULT_SHORTCUTS',
     # Argparse Exports
     'Cmd2ArgumentParser',
+    'TextGroup',
     'register_argparse_argument_parameter',
     'set_default_ap_completer_type',
     'set_default_argument_parser_type',
@@ -68,7 +70,7 @@ __all__: list[str] = [  # noqa: RUF022
     'CommandSet',
     'Statement',
     # Colors
-    "Color",
+    'Color',
     # Completion
     'Choices',
     'CompletionItem',